# docs/swagger.yaml
basePath: /api
definitions:
  httpErrors.ErrResponse:
    properties:
      msg:
        description: application-level error message, for debugging
        type: string
      status:
        description: http response status code
        type: integer
      statusText:
        description: user-level status message
        type: string
    type: object
  presenter.ForgotPassword:
    properties:
      email:
        example: hiennguyen9874@gmail.com
        type: string
    required:
    - email
    type: object
...
  presenter.PublicKey:
    properties:
      public_key_access_token:
        type: string
      public_key_refresh_token:
        type: string
    type: object
  presenter.ResetPassword:
    properties:
      confirm_password:
        example: password
        minLength: 8
        type: string
      new_password:
        example: password
        minLength: 8
        type: string
    required:
    - confirm_password
    - new_password
    type: object
  presenter.Token:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
...
  responses.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/httpErrors.ErrResponse'
      is_success:
        type: boolean
    type: object
info:
  contact: {}
  title: Go boilerplate
  version: "1.0"
paths:
  /auth/forgotpassword:
    post:
      consumes:
      - application/json
      description: Forgot password, code will send to email.
      parameters:
      - description: Forgot Password
        in: body
        name: forgotPassword
        required: true
        schema:
          $ref: '#/definitions/presenter.ForgotPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Forgot password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - multipart/form-data
      description: Sign in, get access token for future requests.
      parameters:
      - description: email
        in: formData
        name: username
        required: true
        type: string
      - description: password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.Token'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Sign In
      tags:
      - auth
  /auth/logout:
    get:
      consumes:
      - application/json
      description: Logout, remove current refresh token in db.
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Logout
      tags:
      - auth
  /auth/logoutall:
    get:
      consumes:
      - application/json
      description: Logout all session of this user.
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Logout all session
      tags:
      - auth
  /auth/publickey:
    get:
      consumes:
      - application/json
      description: Get rsa public key to decode token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.PublicKey'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Get public key
      tags:
      - auth
  /auth/refresh:
    get:
      consumes:
      - application/json
      description: Get new access token from refresh token.
      parameters:
      - description: Authentication header
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/presenter.Token'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Refresh token
      tags:
      - auth
  /auth/resetpassword:
    patch:
      consumes:
      - application/json
      description: Reset Password, using code from email.
      parameters:
      - description: code
        format: code
        in: query
        name: code
        required: true
        type: string
      - description: Reset Password
        in: body
        name: resetPassword
        required: true
        schema:
          $ref: '#/definitions/presenter.ResetPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Reset Password
      tags:
      - auth
  /auth/verifyemail:
    get:
      consumes:
      - application/json
      description: Verify user using code from email.
      parameters:
      - description: offset
        format: code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Response'
      summary: Verify user
      tags:
      - auth
...
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: /api/auth/login
    type: oauth2
swagger: "2.0"